feat(web): combine auth settings
