+++ web/src/routes/admin/system-settings/+page.svelte
@@ -1,34 +1,33 @@
 <script lang="ts">
   import AdminSettings from '$lib/components/admin-page/settings/admin-settings.svelte';
+  import AuthSettings from '$lib/components/admin-page/settings/auth/auth-settings.svelte';
   import FFmpegSettings from '$lib/components/admin-page/settings/ffmpeg/ffmpeg-settings.svelte';
+  import ImageSettings from '$lib/components/admin-page/settings/image/image-settings.svelte';
   import JobSettings from '$lib/components/admin-page/settings/job-settings/job-settings.svelte';
   import LibrarySettings from '$lib/components/admin-page/settings/library-settings/library-settings.svelte';
   import LoggingSettings from '$lib/components/admin-page/settings/logging-settings/logging-settings.svelte';
   import MachineLearningSettings from '$lib/components/admin-page/settings/machine-learning-settings/machine-learning-settings.svelte';
   import MapSettings from '$lib/components/admin-page/settings/map-settings/map-settings.svelte';
   import NewVersionCheckSettings from '$lib/components/admin-page/settings/new-version-check-settings/new-version-check-settings.svelte';
-  import OAuthSettings from '$lib/components/admin-page/settings/oauth/oauth-settings.svelte';
-  import PasswordLoginSettings from '$lib/components/admin-page/settings/password-login/password-login-settings.svelte';
-  import ServerSettings from '$lib/components/admin-page/settings/server/server-settings.svelte';
   import NotificationSettings from '$lib/components/admin-page/settings/notification-settings/notification-settings.svelte';
-  import SettingAccordion from '$lib/components/shared-components/settings/setting-accordion.svelte';
+  import ServerSettings from '$lib/components/admin-page/settings/server/server-settings.svelte';
   import StorageTemplateSettings from '$lib/components/admin-page/settings/storage-template/storage-template-settings.svelte';
   import ThemeSettings from '$lib/components/admin-page/settings/theme/theme-settings.svelte';
-  import ImageSettings from '$lib/components/admin-page/settings/image/image-settings.svelte';
   import TrashSettings from '$lib/components/admin-page/settings/trash-settings/trash-settings.svelte';
   import UserSettings from '$lib/components/admin-page/settings/user-settings/user-settings.svelte';
   import LinkButton from '$lib/components/elements/buttons/link-button.svelte';
   import Icon from '$lib/components/elements/icon.svelte';
   import UserPageLayout from '$lib/components/layouts/user-page-layout.svelte';
+  import SettingAccordionState from '$lib/components/shared-components/settings/setting-accordion-state.svelte';
+  import SettingAccordion from '$lib/components/shared-components/settings/setting-accordion.svelte';
+  import { QueryParameter } from '$lib/constants';
   import { downloadManager } from '$lib/stores/download';
   import { featureFlags } from '$lib/stores/server-config.store';
   import { copyToClipboard } from '$lib/utils';
   import { downloadBlob } from '$lib/utils/asset-utils';
+  import type { SystemConfigDto } from '@immich/sdk';
   import { mdiAlert, mdiContentCopy, mdiDownload, mdiUpload } from '@mdi/js';
   import type { PageData } from './$types';
-  import SettingAccordionState from '$lib/components/shared-components/settings/setting-accordion-state.svelte';
-  import { QueryParameter } from '$lib/constants';
-  import type { SystemConfigDto } from '@immich/sdk';
 
   export let data: PageData;
 
@@ -36,13 +35,12 @@
   let handleSave: (update: Partial<SystemConfigDto>) => Promise<void>;
 
   type Settings =
+    | typeof AuthSettings
     | typeof JobSettings
     | typeof LibrarySettings
     | typeof LoggingSettings
     | typeof MachineLearningSettings
     | typeof MapSettings
-    | typeof OAuthSettings
-    | typeof PasswordLoginSettings
     | typeof ServerSettings
     | typeof StorageTemplateSettings
     | typeof ThemeSettings
@@ -82,6 +80,12 @@
     subtitle: string;
     key: string;
   }> = [
+    {
+      item: AuthSettings,
+      title: 'Authentication Settings',
+      subtitle: 'Manage password, OAuth, and other authentication settings',
+      key: 'image',
+    },
     {
       item: ImageSettings,
       title: 'Image Settings',
@@ -123,18 +127,6 @@
       title: 'Notification Settings',
       subtitle: 'Manage notification settings, including email',
       key: 'notifications',
-    },
-    {
-      item: OAuthSettings,
-      title: 'OAuth Authentication',
-      subtitle: 'Manage the login with OAuth settings',
-      key: 'oauth',
-    },
-    {
-      item: PasswordLoginSettings,
-      title: 'Password Authentication',
-      subtitle: 'Manage the login with password settings',
-      key: 'password',
     },
     {
       item: ServerSettings,
